% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{bnr-class}
\alias{bnr-class}
\title{Bivariate Regression Model}
\description{
Bivariate Regression Model
}
\section{Slots}{

\describe{
\item{\code{Coefficients}}{Regression coefficients.}

\item{\code{Covariance}}{Outcome covariance matrix.}

\item{\code{Information}}{Information.}

\item{\code{Residuals}}{Phenotypic residuals.}
}}

