% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{CheckInit}
\alias{CheckInit}
\title{Check Initiation}
\usage{
CheckInit(init)
}
\arguments{
\item{init}{Optional list of initial parameters for fitting the null model.}
}
\description{
Check Initiation
}
