% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{CheckTestSpec}
\alias{CheckTestSpec}
\title{Check Test Specification}
\usage{
CheckTestSpec(is_zero, p)
}
\arguments{
\item{is_zero}{Logical vector, with as many entires as columns in the target model
matrix, indicating which columns have coefficient zero under the null.}

\item{p}{Number of columns for the target model matrix.}
}
\description{
Check Test Specification
}
