% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{coef.bnr}
\alias{coef.bnr}
\title{Extract Coefficients from Bivariate Regression Model}
\usage{
\method{coef}{bnr}(object, ..., type = NULL)
}
\arguments{
\item{object}{\code{bnr} object.}

\item{...}{Unused.}

\item{type}{Either Target or Surrogate.}
}
\description{
Extract Coefficients from Bivariate Regression Model
}
