% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance-Information.R
\name{covInfo}
\alias{covInfo}
\title{Covariance Information}
\usage{
covInfo(P, S)
}
\arguments{
\item{P}{List of partitioned data. See \code{\link{partSubj}}.}

\item{S}{Target-surrogate covariance matrix.}
}
\value{
Information matrix for the residual target variance (Itt),
  target-surrogate covariance, and surrogate variance.
}
\description{
Covariance Information
}
