% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance-Update.R
\name{covUpdate}
\alias{covUpdate}
\title{Covariate Update}
\usage{
covUpdate(P, b0, a0, b1, a1, S0)
}
\arguments{
\item{P}{List of partitioned data. See \code{\link{partSubj}}.}

\item{b0}{Previous target regression coefficient.}

\item{a0}{Previous surrogate regression coefficient.}

\item{b1}{Current target regression coefficient.}

\item{a1}{Current surrogate regression coefficient.}

\item{S0}{Initial target-surrogate covariance matrix.}
}
\value{
ECM update of the target-surrogate covariance matrix.
}
\description{
Covariate Update
}
