% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{fit.bnr}
\alias{fit.bnr}
\title{Fit Bivariate Normal Regression Model.}
\usage{
fit.bnr(t, s, X, Z = NULL, sig = 0.05, ...)
}
\arguments{
\item{t}{Target outcome vector.}

\item{s}{Surrogate outcome vector.}

\item{X}{Target model matrix.}

\item{Z}{Surrogate model matrix. Defaults to \code{X}.}

\item{sig}{Significance level.}

\item{...}{Additional arguments accepted if fitting via EM.}
}
\value{
An object of class 'mnr' with slots containing the estimated regression
 coefficients, the target-surrogate covariance matrix, the information matrices
 for regression parameters, and the residuals.
}
\description{
Estimation procedure for bivariate normal regression models. The EM algorithm
is applied if \code{s} contains missing values, or if \code{X} differs from
\code{Z}. Otherwise, an accelerated least squares procedure is applied.
}
\details{
The target and surrogate model matrices are expected in numeric format.
Include an intercept if required. Expand factors and interactions in advance.
See \code{\link{fit.bnem}} for additional options accepted if fitting via EM.
}
\examples{
\dontrun{
# See `? rBNR` for data generation
fit.bnem(t=t,s=s,X=X,Z=Z,maxit=100,eps=1e-8,report=T);
}
}
