% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{FormatOutput}
\alias{FormatOutput}
\title{Format Output}
\usage{
FormatOutput(data_part, method, b, a, sigma, sig)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{method}{Estimation method.}

\item{b}{Final target regression parameter.}

\item{a}{Final surrogate regression parameter.}

\item{sigma}{Final target-surrogate covariance matrix.}

\item{sig}{Significance level.}
}
\value{
Object of class 'bnr'.
}
\description{
Format Output
}
