% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation-EM.R
\name{IterUpdate}
\alias{IterUpdate}
\title{Update Iteration}
\usage{
IterUpdate(theta0, update, maxit, eps, report)
}
\arguments{
\item{theta0}{List containing the initial parameter values.}

\item{update}{Function to iterate. Should accept and return a list similar
parameter values.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report fitting progress?}
}
\description{
Update Iteration
}
