% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linearPredictors}
\alias{linearPredictors}
\title{Linear predictors for bivariate regression}
\usage{
linearPredictors(X, Z, b, a)
}
\arguments{
\item{X}{Target design matrix (N x q).}

\item{Z}{Surrogate design matrix (N x r).}

\item{b}{Target coefficient (q x 1).}

\item{a}{Surrogate coefficient (r x 1).}
}
\value{
Nx2 matrix with columns \eqn{X\beta} and \eqn{Z\alpha}.
}
\description{
Computes the Nx2 matrix of linear predictors \eqn{[X\beta, Z\alpha]}.
}
