% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matDet}
\alias{matDet}
\title{Matrix Determinant}
\usage{
matDet(A, logDet = FALSE)
}
\arguments{
\item{A}{Numeric matrix.}

\item{logDet}{Return the logarithm of the determinant?}
}
\value{
Scalar.
}
\description{
Calculates the determinant of \eqn{A}.
}
