% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matInv}
\alias{matInv}
\title{Matrix Inverse}
\usage{
matInv(A)
}
\arguments{
\item{A}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
Calculates \eqn{A^{-1}}. Uses a direct solve for small matrices (3x3 or
smaller) and pseudo-inverse otherwise.
}
