% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matOP}
\alias{matOP}
\title{Matrix Outer Product}
\usage{
matOP(A, B)
}
\arguments{
\item{A}{Numeric matrix.}

\item{B}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
Calculates the outer product \eqn{AB'}.
}
