% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observed-LogLikelihood.R
\name{obsLogLik}
\alias{obsLogLik}
\title{Observed Data Log Likelihood}
\usage{
obsLogLik(P, b, a, S)
}
\arguments{
\item{P}{List of partitioned data. See \code{\link{partSubj}}.}

\item{b}{Target regression coefficient.}

\item{a}{Surrogate regression coefficient}

\item{S}{Target-surrogate covariance}
}
\value{
Observed data log likelihood.
}
\description{
Observed Data Log Likelihood
}
