% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit-EM.R
\name{paramInit}
\alias{paramInit}
\title{Parameter Initialization}
\usage{
paramInit(P, b0, a0, S0)
}
\arguments{
\item{P}{List of partitioned data. See \code{\link{partSubj}}.}

\item{b0}{Initial target regression coefficient.}

\item{a0}{Initial surrogate regression coefficient.}

\item{S0}{Initial covariance matrix.}
}
\value{
List containing initial values of beta, alpha, sigma.
}
\description{
Parameter Initialization
}
