% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partition-Subjects.R
\name{partSubj}
\alias{partSubj}
\title{Partition Data by Missingness}
\usage{
partSubj(t, s, X, Z = NULL)
}
\arguments{
\item{t}{Target outcome vector.}

\item{s}{Surrogate outcome vector.}

\item{X}{Target model matrix.}

\item{Z}{Surrogate model matrix.}
}
\value{
List containing the partitioned data.
}
\description{
Partition Data by Missingness
}
