% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.bnr}
\alias{print.bnr}
\title{Print for Bivariate Regression Model}
\usage{
\method{print}{bnr}(x, ..., type = "Regression")
}
\arguments{
\item{x}{\code{bnr} object.}

\item{...}{Unused.}

\item{type}{Either Regression or Covariance.}
}
\description{
Print for Bivariate Regression Model
}
