% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rBNR}
\alias{rBNR}
\title{Simulate Bivariate Normal Data with Missingness}
\usage{
rBNR(X, Z, b, a, mt = 0, ms = 0, S)
}
\arguments{
\item{X}{Target design matrix.}

\item{Z}{Surrogate design matrix.}

\item{b}{Target regression coefficient.}

\item{a}{Surrogate regression coefficient.}

\item{mt}{Target missingness in [0,1].}

\item{ms}{Surrogate missingness in [0,1].}

\item{S}{Target-surrogate covariance structure.}
}
\value{
Numeric \eqn{n} by 2 matrix. The first column contains the target
  outcome, the second contains the surrogate outcome.
}
\description{
Function to simulate from a bivariate normal regression model with outcomes
missing completely at random.
}
\examples{
\dontrun{
set.seed(100);
# Observations
n = 1e3;
# Target design
X = cbind(1,matrix(rnorm(3*n),nrow=n));
# Surrogate design
Z = cbind(1,matrix(rnorm(3*n),nrow=n));
# Target coefficient
b = c(-1,0.1,-0.1,0.1);
# Surrogate coefficient
a = c(1,-0.1,0.1,-0.1);
# Covariance structure
S = matrix(c(1,0.5,0.5,1),nrow=2);
# Data generation, target and surrogate subject to 10\% missingness
Y = rBNR(X,Z,b,a,mt=0.1,ms=0.1,S=S);
}
}
