% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression-Information.R
\name{regInfo}
\alias{regInfo}
\title{Regression Information}
\usage{
regInfo(P, S)
}
\arguments{
\item{P}{List of partitioned data. See \code{\link{partSubj}}.}

\item{S}{Target-surrogate covariance matrix.}
}
\value{
List containing the information matrix for beta (Ibb), the
  information matrix for alpha (Iaa), and the cross information (Iba).
}
\description{
Regression Information
}
