% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabulate.R
\name{regTab}
\alias{regTab}
\title{Tabulate Regression Coefficients}
\usage{
regTab(point, info, sig = 0.05)
}
\arguments{
\item{point}{Point estimates.}

\item{info}{Information matrix.}

\item{sig}{Significance level.}
}
\value{
Data.table containing the point estimate, standard error, confidence
  interval, and Wald p-value.
}
\description{
Tabulate Regression Coefficients
}
