% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression-Update.R
\name{regUpdate}
\alias{regUpdate}
\title{Regression Update}
\usage{
regUpdate(P, S)
}
\arguments{
\item{P}{List of partitioned data. See \code{\link{partSubj}}.}

\item{S}{Target-surrogate covariance matrix.}
}
\value{
List containing the generalized least squares estimates of beta and
  alpha.
}
\description{
Regression Update
}
