% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{residuals.bnr}
\alias{residuals.bnr}
\title{Extract Residuals from Bivariate Regression Model}
\usage{
\method{residuals}{bnr}(object, ..., type = NULL)
}
\arguments{
\item{object}{A \code{bnr} object.}

\item{...}{Unused.}

\item{type}{Either Target or Surrogate.}
}
\description{
Extract Residuals from Bivariate Regression Model
}
