% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SchurC}
\alias{SchurC}
\title{Schur complement}
\usage{
SchurC(Ibb, Iaa, Iba)
}
\arguments{
\item{Ibb}{Information of target parameter}

\item{Iaa}{Information of nuisance parameter}

\item{Iba}{Cross information between target and nuisance parameters}
}
\value{
Numeric matrix.
}
\description{
Calculates the efficient information \eqn{I_{bb}-I_{ba}I_{aa}^{-1}I_{ab}}.
}
