% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_LS.R
\name{Score.bnls}
\alias{Score.bnls}
\title{Score Test via Least Squares.}
\usage{
Score.bnls(t, s, X, L)
}
\arguments{
\item{t}{Target outcome vector.}

\item{s}{Surrogate outcome vector.}

\item{X}{Model matrix.}

\item{L}{Logical vector, with as many entires as columns in the target model
matrix, indicating which columns have coefficient zero under the null.}
}
\value{
A numeric vector containing the score statistic, the degrees of
  freedom, and a p-value.
}
\description{
Performs a Score test of the null hypothesis that a subset of the regression
parameters for the target outcome are zero.
}
