% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spray-package.R
\docType{package}
\name{Spray}
\alias{Spray}
\alias{Spray-package}
\title{Spray: Surrogate Phenotype Regression Analysis}
\description{
This package performs estimation and inference for the parameters of a
bivariate normal regression model in which elements of the response matrix
are missing at random. In the case of bilateral missingness, estimation is
performed using \code{\link{fit.bnem}}. In the case of unilaterial
missingness, estimation is performed using \code{\link{fit.bnls}}. Inference
on regression parameters for the target outcome is performed using
\code{\link{test.bnr}}.
}
\author{
Zachary R. McCaw
}
