% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation-EM.R
\name{UpdateEM}
\alias{UpdateEM}
\title{EM Update}
\usage{
UpdateEM(data_part, b0, a0, sigma0)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{b0}{Initial target regression coefficient.}

\item{a0}{Initial surrogate regression coefficient.}

\item{sigma0}{Initial covariance matrix.}
}
\value{
List containing updated values for beta 'b', alpha 'a', 'sigma', the
  log likelihood 'loglik', and the change in log likelihood 'delta'.
}
\description{
EM Update
}
