% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{vcov.bnr}
\alias{vcov.bnr}
\title{Extract Covariance Matrix from Multivariate Regression Model}
\usage{
\method{vcov}{bnr}(object, ..., type = "Information", inv = F)
}
\arguments{
\item{object}{A \code{bnr} object.}

\item{...}{Unused.}

\item{type}{Either Outcome or Information. Default is Information.}

\item{inv}{Invert the covariance matrix? Default is FALSE.}
}
\description{
Returns the either the estimated covariance matrix of the outcome,
or the information matrix of the regression coefficients.
}
