% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald-LS.R
\name{WaldBNLS}
\alias{WaldBNLS}
\title{Wald Test via Least Squares.}
\usage{
WaldBNLS(t, s, X, is_zero)
}
\arguments{
\item{t}{Target outcome vector.}

\item{s}{Surrogate outcome vector.}

\item{X}{Model matrix.}

\item{is_zero}{Logical vector, with as many entires as columns in the target model
matrix, indicating which columns have coefficient zero under the null.}
}
\value{
A numeric vector containing the Wald statistic, the degrees of
  freedom, and a p-value.
}
\description{
Performs a Wald test of the null hypothesis that a subset of the regression
parameters for the target outcome are zero.
}
