% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{bnr-class}
\alias{bnr-class}
\title{Bivariate Regression Model}
\description{
Bivariate Regression Model
}
\section{Slots}{

\describe{
\item{\code{Covariance}}{Residual covariance matrix.}

\item{\code{Covariance.info}}{Information for covariance parameters.}

\item{\code{Covariance.tab}}{Table of covariance parameters.}

\item{\code{Regression.info}}{Information for regression coefficients.}

\item{\code{Regression.tab}}{Table of regression coefficients.}

\item{\code{Residuals}}{Outcome residuals.}
}}

