% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,bnr-method}
\alias{show,bnr-method}
\title{Show for Bivariate Regression Model}
\usage{
\S4method{show}{bnr}(object)
}
\arguments{
\item{object}{\code{bnr} object.}
}
\description{
Show for Bivariate Regression Model
}
