% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information-Covariance.R
\name{CovInfo}
\alias{CovInfo}
\title{Covariance Information Matrix}
\usage{
CovInfo(data_part, sigma)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{sigma}{Target-surrogate covariance matrix.}
}
\value{
3x3 Numeric information matrix for the target variance,
  target-surrogate covariance, and surrogate variance.
}
\description{
Covariance Information Matrix
}
