% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{CovTab}
\alias{CovTab}
\title{Tabulate Covariance Parameters}
\usage{
CovTab(point, info, sig = 0.05)
}
\arguments{
\item{point}{Point estimates.}

\item{info}{Information matrix.}

\item{sig}{Significance level.}
}
\value{
Data.table containing the point estimate, standard error, and
  confidence interval.
}
\description{
Tabulate Covariance Parameters
}
