% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update-Covariance.R
\name{CovUpdate}
\alias{CovUpdate}
\title{Covariate Update}
\usage{
CovUpdate(data_part, b0, a0, b1, a1, sigma0)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{b0}{Previous target regression coefficient.}

\item{a0}{Previous surrogate regression coefficient.}

\item{b1}{Current target regression coefficient.}

\item{a1}{Current surrogate regression coefficient.}

\item{sigma0}{Initial target-surrogate covariance matrix.}
}
\value{
ECM update of the target-surrogate covariance matrix.
}
\description{
Covariate Update
}
