% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit-LS.R
\name{fit.bnls}
\alias{fit.bnls}
\title{Fit Bivariate Normal Regression Model via Least Squares}
\usage{
fit.bnls(t, s, X, sig = 0.05)
}
\arguments{
\item{t}{Target outcome vector.}

\item{s}{Surrogate outcome vector.}

\item{X}{Model matrix.}

\item{sig}{Significance level.}
}
\value{
An object of class 'bnr' with slots containing the estimated
  regression coefficients, the target-surrogate covariance matrix, the
  information matrices for the regression and covariance parameters, and the
  residuals.
}
\description{
Estimation procedure for bivariate normal regression models in which
only the target outcome is subject to missingness.
}
\details{
The model matrix is expected in numeric format. Include an intercept if
required. Expand factors and interactions in advance.
}
\examples{
\dontrun{
# See `? rBNR` for data generation
fit.bnls(t=t,s=s,X=X);
}
}
