% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikelihood.R
\name{ObsLogLik}
\alias{ObsLogLik}
\title{Observed Data Log Likelihood}
\usage{
ObsLogLik(data_part, b, a, sigma)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{b}{Target regression coefficient.}

\item{a}{Surrogate regression coefficient.}

\item{sigma}{Target-surrogate covariance matrix.}
}
\value{
Observed data log likelihood.
}
\description{
Observed Data Log Likelihood
}
