% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation-EM.R
\name{ParamInit}
\alias{ParamInit}
\title{Parameter Initialization}
\usage{
ParamInit(data_part, b0, a0, sigma0)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{b0}{Initial target regression coefficient.}

\item{a0}{Initial surrogate regression coefficient.}

\item{sigma0}{Initial covariance matrix.}
}
\value{
List containing initial values of beta, alpha, sigma.
}
\description{
Parameter Initialization
}
