% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information-Regression.R
\name{RegInfo}
\alias{RegInfo}
\title{Regression Information}
\usage{
RegInfo(data_part, sigma, as_matrix = FALSE)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{sigma}{Target-surrogate covariance matrix.}

\item{as_matrix}{Return as an information matrix? If FALSE, returns a list.}
}
\value{
List containing the information matrix for beta (Ibb), the
  information matrix for alpha (Iaa), and the cross information (Iba).
}
\description{
Regression Information
}
