% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update-Regression.R
\name{RegUpdate}
\alias{RegUpdate}
\title{Regression Update}
\usage{
RegUpdate(data_part, sigma)
}
\arguments{
\item{data_part}{List of partitioned data. See \code{\link{PartitionData}}.}

\item{sigma}{Target-surrogate covariance matrix.}
}
\value{
List containing the generalized least squares estimates of beta and
  alpha.
}
\description{
Regression Update
}
