% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spray-package.R
\docType{package}
\name{Spray}
\alias{Spray}
\title{Spray: Surrogate Phenotype Regression Analysis}
\description{
This package performs estimation and inference for the parameters of a
bivariate normal regression model in which elements of the response matrix
are missing at random. The primary estimation function is
\code{\link{Fit.BNR}}. In the case of bilateral missingness, estimation is
performed using \code{\link{Fit.BNEM}}. In the case of unilaterial
missingness, estimation is performed using \code{\link{Fit.BNLS}}. Inference
on regression parameters for the target outcome is performed using
\code{\link{Test.BNR}}.
}
\author{
Zachary R. McCaw
}
