% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.R
\name{test.bnr}
\alias{test.bnr}
\title{Test Bivariate Normal Regression Model.}
\usage{
test.bnr(t, s, X, Z = NULL, L, test = "Wald", ...)
}
\arguments{
\item{t}{Target outcome vector.}

\item{s}{Surrogate outcome vector.}

\item{X}{Model matrix.}

\item{Z}{Surrogate model matrix.}

\item{L}{Logical vector, with as many entires as columns in the target model
matrix, indicating which columns have coefficient zero under the null.}

\item{test}{Either Score or Wald. Only Wald is available with LS.}

\item{...}{Additional arguments accepted if fitting via EM.}
}
\value{
A numeric vector containing the test statistic, the degrees of
  freedom, and a p-value.
}
\description{
Performs a test of the null hypothesis that a subset of the regression
parameters for the target outcome are zero in the bivariate normal regression
model.
}
\examples{
\dontrun{
# See `? rBNR` for data generation
# See vignette for test description
test.bnr(t=t,s=s,X=X,L=c(F,T,T,T));
test.bnr(t=t,s=s,X=X,L=c(F,T,T,F));
test.bnr(t=t,s=s,X=X,L=c(F,F,F,T));
}
}
