% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{vcov.bnr}
\alias{vcov.bnr}
\title{Extract Covariance Matrix from Bivariate Normal Regression Model}
\usage{
\method{vcov}{bnr}(object, ..., type = "Regression", inv = FALSE)
}
\arguments{
\item{object}{\code{bnr} object.}

\item{...}{Unused.}

\item{type}{Select "Covariance","Outcome",or "Regression". Default is
"Regression".}

\item{inv}{Invert the covariance matrix? Default is FALSE.}
}
\description{
Returns the either the estimated covariance matrix of the outcome, the
information matrix for regression coefficients, or the information matrix for
covariance parameters.
}
